requirements_final.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# OCR
pytesseract==0.3.10
easyocr==1.7.0

# Image/PDF handling
pdf2image==1.16.3
PyMuPDF==1.23.7
pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.2

# Validation / Pydantic
pydantic==2.5.0

# Optional but recommended (for file ops)
python-dotenv==1.0.0
